Show stacktrace, Split instruction executor
